Make FlourSelection a Component containing set of buttons as a subcomponent
